convert individual event information to aggregated information per network node

convert individual event information to aggregated information per network node

aggr_data(dat, from = NULL, cumsum = TRUE)

Arguments

dat

data.frame with variables 'node', 'time', 'delay', events data with single events with count magnitude

from

character in strftime format, e.g. "2014-06-12 16:15", data is subsetted accordingly before aggregation

cumsum

logical indicating whether data is aggregated by cumulative sum, default is TRUE

Value

data.frame of dimension (TxK), where T is the number of observation times and K the number of network nodes. Thus, each row represents a snapshot of the spreading process at a specific observation time with the event magnitude observed at the network nodes. Rownames are observation times, colnames are node names.

See also