analyze public transportation network characteristics

analyze public transportation network characteristics

analyze_ptn(g)

Arguments

g

igraph object, network graph representing the public transportation network, vertices represent stations, which are linked by an edge if there is a direct transfer between them

Value

'data.frame': 1 obs. of 7 variables:

vcount number of nodes,

ecount number of edges,

density network graph density,

av_deg average degree,

av_cent average unit betweenness,

diam diameter, and

trans transitivity.

References

Details to the computation and interpretation can be found in:

Kolaczyk, E. D. (2009). Statistical analysis of network data: methods and models. Springer series in statistics. Springer. <DOI: 10.1007/978-0-387-88146-1>

Manitz, J. (2014): Statistical Inference for Propagation Processes on Complex Networks. Ph.D. thesis, Georg-August-University Goettingen. Verlag Dr.~Hut, ISBN 978-3-8439-1668-4. Available online: https://ediss.uni-goettingen.de/handle/11858/00-1735-0000-0022-5F38-B.

See also

Examples

data(ptnAth)
analyze_ptn(ptnAth)
#>   vcount ecount    density   av_deg   av_cent   av_spl diam trans
#> 1     51     52 0.04078431 2.039216 0.1769348 9.669804   29     0

data(ptnGoe)
analyze_ptn(ptnGoe)
#>   vcount ecount    density   av_deg    av_cent   av_spl diam     trans
#> 1    257    548 0.00832928 4.264591 0.09582916 13.21822   35 0.1189024